// template/react-starter-kit/src/.generated/api/model/inquiryStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type InquiryStatus = typeof InquiryStatus[keyof typeof InquiryStatus];


export const InquiryStatus = {
  pending: 'pending',
  answered: 'answered',
  closed: 'closed',
} as const;